Solutions day 15.
Decided on doing it object oriented. Dont think it was necessary, but it was to early to do it without.

For part 2, i had to copy the grid 25 times, but with adding values. I may have done this a little obnoxious, but in the end is fine.

I also read the input file again because i split the code in 2 methods.

These methods do contain duplicate code, so in the end i should refactor it so that it doesnt contain double code anymore.