I made sure that the duplicated code is called in methods, and the code looks a bit cleaner.
It only had a few 100 ms effect, and it still takes about 5 seconds to complete the second part. But that is maybe just how cumbersome i implemented it. (from input file to grid to a node map.

Maybe I also fucked up with the implementation of Dijkstra. Its working, but I probably iterate over something a bit too much.